// frontend/types/supplier.ts
export type Supplier = {
  sup_id: string; // ✅ KEEP THIS STRING ONLY
  sup_name?: string;
  sup_add?: string;
  sup_person?: string;
  sup_phone?: string;
  sup_email?: string;
  sup_gst?: string;
  acct_no?: string;
  acct_name?: string;
  bank_name?: string;
  ifsc_code?: string;
  bank_branch?: string;
};